+++ src/runtime/components/prose/AccordionItem.vue
@@ -1,0 +1,38 @@
+<script lang="ts">
+import type { AppConfig } from '@nuxt/schema'
+import type { ComponentConfig } from '../../types/tv'
+import theme from '#build/b24ui/prose/accordion-item'
+
+type ProseAccordionItem = ComponentConfig<typeof theme, AppConfig, 'accordionItem', 'b24ui.prose'>
+
+export interface ProseAccordionItemProps {
+  label: string
+  description?: string
+  class?: any
+}
+
+export interface ProseAccordionItemSlots {
+  default(props?: {}): any
+}
+</script>
+
+<script setup lang="ts">
+import { computed } from 'vue'
+import { useAppConfig } from '#imports'
+import { tv } from '../../utils/tv'
+
+const props = defineProps<ProseAccordionItemProps>()
+defineSlots<ProseAccordionItemSlots>()
+
+const appConfig = useAppConfig() as ProseAccordionItem['AppConfig']
+
+const b24ui = computed(() => tv({ extend: tv(theme), ...(appConfig.b24ui?.prose?.accordionItem || {}) }))
+</script>
+
+<template>
+  <div :class="b24ui({ class: props.class })">
+    <slot>
+      {{ description }}
+    </slot>
+  </div>
+</template>