+++ src/theme/prose/accordion-item.ts
@@ -1,0 +1,8 @@
+/**
+ * ProseAccordionItem
+ * ---
+ */
+
+export default {
+  base: 'pb-[20px] text-(--b24ui-typography-description-color) *:first:mt-0 *:last:mb-0 *:my-[6px]'
+}