+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * feat(ProseTabs): new component
 * feat(ProseTabsItem): new component
 * feat(ProseAccordion): new component
+* feat(ProseAccordionItem): new component
 * feat(ProseKbd): new component
 * feat(Banner): new component
 * feat(Card): new component